delete old modules, updates, fixes
