--- engine-core/src/constants.rs
@@ -1,13 +1,0 @@
-// This file contains constants used in the engine library
-
-pub const DB_EVENT_LOG_ERROR: &str = "Error occurred while trying to log database event";
-pub const DB_NOT_FOUND: &str = "Database was not found";
-pub const DB_EXISTS: &str = "Database already exists";
-pub const COLLECTION_NOT_FOUND: &str = "Collection was not found";
-pub const DOCUMENT_NOT_FOUND: &str = "Document was not found";
-
-// Path to databases directory in filesystem
-pub const DB_DIR_PATH: &str = "./databases";
-
-// Database file extension
-// pub const DB_FILE_EXTENSION: &str = "json";
