+++ engine-core/src/lib.rs
@@ -6,16 +6,15 @@ pub mod logging;
 pub mod db;
 mod db_manager_pb;
 mod input_data;
-pub mod constants;
 mod api;
 
 use std::path::PathBuf;
 pub use api::EngineApi;
-use constants::DB_DIR_PATH;
 pub use serde_json;
 pub use db_manager_pb::DatabaseManager;
 pub use input_data::DocumentInputDataField;
 
+const DB_DIR_PATH: &str = "./databases";
 // Engine version
 const VERSION: &str = env!("CARGO_PKG_VERSION");
